requirements.txt
amqp==2.5.2
backcall==0.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
bs4==0.0.1
celery==4.4.2
certifi==2020.4.5.1
chardet==3.0.4
decorator==4.4.2
Django==2.2.12
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-extensions==2.2.9
django-filter==2.2.0
djangorestframework==3.11.0
Faker==4.1.0
flake8==3.8.2
flake8-builtins==1.5.3
flake8-import-order==0.18.1
flake8-print==3.1.4
idna==2.9
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
jedi==0.17.0
kombu==4.6.8
lxml==4.5.1
Markdown==3.2.2
mccabe==0.6.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
pkg-resources==0.0.0
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
Pygments==2.6.1
python-dateutil==2.8.1
pytz==2020.1
requests==2.23.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
text-unidecode==1.3
traitlets==4.3.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
XlsxWriter==1.2.9
zipp==3.1.0
